start refactoring into modules
