+++ kafka-junit5/src/main/java/com/salesforce/kafka/test/junit5/KafkaTestServer.java
@@ -23,7 +23,7 @@
  * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.salesforce.kafka.test;
+package com.salesforce.kafka.test.junit5;
 
 import com.google.common.collect.Maps;
 import com.google.common.io.Files;
