+++ kafka-junit5/src/main/java/com/salesforce/kafka/test/junit5/SharedKafkaTestResource.java
@@ -23,10 +23,8 @@
  * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.salesforce.kafka.test.junit;
+package com.salesforce.kafka.test.junit5;
 
-import com.salesforce.kafka.test.KafkaTestServer;
-import com.salesforce.kafka.test.KafkaTestUtils;
 import org.apache.curator.test.TestingServer;
 
 /**
