+++ kafka-junit5/src/main/java/com/salesforce/kafka/test/junit5/ZookeeperResourceExtension.java
@@ -23,8 +23,9 @@
  * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.salesforce.kafka.test.junit;
+package com.salesforce.kafka.test.junit5;
 
+import org.apache.curator.test.TestingServer;
 import org.junit.jupiter.api.extension.AfterAllCallback;
 import org.junit.jupiter.api.extension.BeforeAllCallback;
 import org.junit.jupiter.api.extension.ExtensionContext;
@@ -34,60 +35,67 @@ import org.junit.jupiter.api.extension.ParameterResolver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.IOException;
+
 /**
- * JUnit 5 extension to provide an internal test kafka server to be shared across test cases within the same test class.
+ * JUnit 5 extension to provide an internal test zookeeper server to be shared across test cases within the same test class.
  *
  * Annotate your test class with:
- *   &#064;ExtendWith(KafkaResourceExtension.class)
+ *   &#064;ExtendWith(ZookeeperResourceExtension.class)
  *
  * Add a constructor parameter to your test class:
  *
- *   public YourTestClass(SharedKafkaTestResource sharedKafkaTestResource) {
+ *   public YourTestClass(SharedZookeeperTestResource sharedZookeeperTestResource) {
  *     // Save reference to test resource.
- *     this.sharedKafkaTestResource = sharedKafkaTestResource;
+ *     this.sharedZookeeperTestResource = sharedZookeeperTestResource;
  *   }
  *
  * Within your test case methods:
- *   this.sharedKafkaTestResource.getKafkaTestServer()...
+ *   this.sharedZookeeperTestResource.getZookeeperTestServer()...
+ *   this.sharedZookeeperTestResource.getZookeeperConnectString()...
  */
-public class KafkaResourceExtension implements BeforeAllCallback, AfterAllCallback, ParameterResolver {
-    private static final Logger logger = LoggerFactory.getLogger(KafkaResourceExtension.class);
+public class ZookeeperResourceExtension implements BeforeAllCallback, AfterAllCallback, ParameterResolver {
+    private static final Logger logger = LoggerFactory.getLogger(ZookeeperResourceExtension.class);
+
+    private SharedZookeeperTestResource zookeeperTestResource = null;
 
     /**
-     * Shared Kafka Test Server Resource.
+     * Here we shut down the internal test zookeeper service.
      */
-    private SharedKafkaTestResource kafkaTestResource = new SharedKafkaTestResource();
+    @Override
+    public void afterAll(ExtensionContext context) throws Exception {
+        logger.info("Shutting down zookeeper test server");
+
+        // If we don't have an instance
+        if (zookeeperTestResource == null) {
+            // Nothing to close.
+            return;
+        }
+
+        try {
+            final TestingServer testingServer = zookeeperTestResource.getZookeeperTestServer();
+            testingServer.stop();
+            testingServer.close();
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
+        // null out reference
+        zookeeperTestResource = null;
+    }
 
     /**
-     * Here we stand up an internal test kafka and zookeeper service.
+     * Here we stand up an internal test zookeeper service.
      * Once for all tests that use this shared resource.
      */
     @Override
     public void beforeAll(ExtensionContext context) throws Exception {
-        logger.info("Starting kafka test server");
-
-        // Start kafka test server
-        kafkaTestResource
-            .getKafkaTestServer()
-            .start();
-    }
-
-    /**
-     * Here we shut down the internal test kafka and zookeeper services.
-     */
-    @Override
-    public void afterAll(ExtensionContext context) throws Exception {
-        logger.info("Shutting down kafka test server");
-
-        // Close out kafka test server if needed
-        try {
-            kafkaTestResource
-                .getKafkaTestServer()
-                .shutdown();
-        } catch (final Exception e) {
-            throw new RuntimeException(e);
+        logger.info("Starting Zookeeper test server");
+        if (zookeeperTestResource != null) {
+            throw new IllegalStateException("Unknown State! Zookeeper test server already exists!");
         }
-        kafkaTestResource = null;
+        // Setup zookeeper test server
+        zookeeperTestResource = new SharedZookeeperTestResource();
     }
 
     @Override
@@ -97,7 +105,7 @@ public class KafkaResourceExtension implements BeforeAllCallback, AfterAllCallba
         return parameterContext
             .getParameter()
             .getType()
-            .equals(SharedKafkaTestResource.class);
+            .equals(SharedZookeeperTestResource.class);
     }
 
     @Override
@@ -108,8 +116,8 @@ public class KafkaResourceExtension implements BeforeAllCallback, AfterAllCallba
             .getParameter()
             .getType();
 
-        if (parameterType.equals(SharedKafkaTestResource.class)) {
-            return kafkaTestResource;
+        if (parameterType.equals(SharedZookeeperTestResource.class)) {
+            return zookeeperTestResource;
         }
         return null;
     }
