+++ kafka-junit5/src/test/java/com/salesforce/kafka/test/junit5/KafkaTestServerTest.java
@@ -23,11 +23,9 @@
  * USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package com.salesforce.kafka.test;
+package com.salesforce.kafka.test.junit5;
 
 import com.google.common.collect.Lists;
-import com.salesforce.kafka.test.junit.KafkaResourceExtension;
-import com.salesforce.kafka.test.junit.SharedKafkaTestResource;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.ConsumerRecords;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
